Pass arguments to program
Parse and pass the arguments given for program through stack
Pintos had the assumption that programs will be run with no arguments,
so it used the command for file name. Which is not a correct assumption now.
Extract the actual file name from command line for pintos (used for thread name, program file name, etc)