Fix multi-oom test
stop using a different page for every file descriptor. use malloc instead.
free the memory used by file descriptors on thread exit.
add a semaphor to thread to be able to wait for it to load.
if a thread's load fails, return TID_ERROR to the called exec.
free the memory used by thread on thread wait.
add a thread_create_get_thread which returns the created thread,
which then can be used for waiting to load.