Make filesystem files growable
File headers now support containing 123 direct sectors, 128 indirect ones and 128^2 doubly-indirect sectors.
With this change we can grow files on demand! With every write that goes beyond a file's length, try growing it first.
But the growing and creation part got a little bit complicated. Maybe refactor them later?
Also if something goes wrong in the process of growing a file, the allocated sectors will not be free after that. Maybe fix it if needed?

All growing tests passed. Waiting to implement directory to take a look at persistence tests.